Respect node NetworkUnavailable condition
This commit adds the feature of respecting the node NetworkUnavailable condition.
In other words, when a node's NetworkUnavailable condition is True, do not
announce the LB service from the speaker in this node.

This makes the speaker to reload the service handler every time
a node is updated, and adds the logic to delete the LB service announcement
if the NetworkUnavailable condition is True.

Note:
1) We propagate the speaker's node collection into the ShouldAnnounce so
that the layer2_controller can use that as part of the leader election
and have the other nodes be aware of the unavailable node not being part
of the elections.

2) We make the node controller to listen for all node events,
not only to the speaker's node. that is for the L2 re-elections.

3) As for the BGP controller: in the ShouldAnnounce func added the
check for NodeNetworkUnavailable condition, in case of true,
do not announce.

Signed-off-by: liornoy <lnoy@redhat.com>